refactor(userbodymetrics): improve organization with create package
- Also remove measuredAt from update domain method
- Fixes tests to reflect new package
- Remove navy body percentage calculator from create use case
