+++ src/main/java/com/ironcore/interfaces/rest/userbodymetrics/UserBodyMetricsController.java
@@ -1,7 +1,7 @@
 package com.ironcore.interfaces.rest.userbodymetrics;
 
-import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsCommand;
-import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsResult;
+import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsCommand;
+import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsResult;
 import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsUseCase;
 import com.ironcore.infrastructure.security.auth.AuthenticatedUser;
 import com.ironcore.interfaces.rest.userbodymetrics.dto.CreateUserBodyMetricsRequest;
