+++ src/test/java/com/ironcore/application/userbodymetrics/usecase/CreateUserBodyMetricsUseCaseTest.java
@@ -5,8 +5,8 @@ import com.ironcore.application.exception.ResourceNotFoundException;
 import com.ironcore.application.exception.UserInactiveException;
 import com.ironcore.application.logging.audit.payload.LoggableData;
 import com.ironcore.application.logging.audit.port.AuditLogPublisher;
-import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsCommand;
-import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsResult;
+import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsCommand;
+import com.ironcore.application.userbodymetrics.create.CreateUserBodyMetricsResult;
 import com.ironcore.application.userbodymetrics.CreateUserBodyMetricsUseCase;
 import com.ironcore.application.userbodymetrics.UserBodyMetricsAuditData;
 import com.ironcore.domain.logging.audit.enums.AuditActionType;
